requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
Babel==2.10.3
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.1
certifi==2022.9.24
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
cloud-init==23.3.1
coloredlogs==15.0.1
comm==0.1.4
configobj==5.0.8
cryptography==38.0.4
dbus-python==1.3.2
debugpy==1.8.0
decorator==5.1.1
discord.py==2.3.2
distro-info==1.5
dnspython==2.4.2
executing==1.2.0
Flask==2.3.3
frozenlist==1.4.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.60.0
httplib2==0.20.4
humanfriendly==10.0
idna==3.3
iotop==0.6
ipykernel==6.25.2
ipython==8.15.0
iso8601==2.0.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.10.3
jupyter_client==8.3.1
jupyter_core==5.3.1
kali-tweaks==2023.3.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
multidict==6.0.4
mutagen==1.47.0
nest-asyncio==1.5.7
netifaces==0.11.0
oauthlib==3.2.2
openai==0.27.9
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==4.24.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycurl==7.45.2
Pygments==2.16.1
PyGObject==3.46.0
PyJWT==2.7.0
pymongo==4.5.0
PyNaCl==1.5.0
pyparsing==3.1.1
pyrsistent==0.18.1
pyserial==3.5
PySimpleSOAP==1.16.2
python-apt==2.6.0
python-dateutil==2.8.2
python-debian==0.1.49
python-debianbts==4.0.1
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
reportbug==12.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3987==1.3.8
rsa==4.9
schedule==1.2.0
six==1.16.0
stack-data==0.6.2
tabulate==0.9.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
tweepy==4.14.0
twitchio==2.7.0
typing_extensions==4.7.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.16
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.11.1
Werkzeug==2.3.7
yarl==1.9.2
